add department implement

+++ src/main/java/com/worksap/stm2016/utils/CommonUtils.java
@@ -5,11 +5,11 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 
 public class CommonUtils {
 	
-	final static public Integer ROLE_HR_MANAGER=1;
-	final static public Integer ROLE_RECRUITER=2;
-	final static public Integer ROLE_CB_SPECIALIST=3;
-	final static public Integer ROLE_TEAM_MANAGER=4;
-	final static public Integer ROLE_SHORT_TERM_EMPLOYEE=5;
+	final static public Long ROLE_HR_MANAGER=(long) 1;
+	final static public Long ROLE_RECRUITER=(long) 2;
+	final static public Long ROLE_CB_SPECIALIST=(long) 3;
+	final static public Long ROLE_TEAM_MANAGER=(long) 4;
+	final static public Long ROLE_SHORT_TERM_EMPLOYEE=(long) 5;
 	
 	static public PasswordEncoder passwordEncoder() {
 		PasswordEncoder encoder = (PasswordEncoder) new BCryptPasswordEncoder();
